# client/src/renderer/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Octaris Bioprinter",
    "version": "0.1.0"
  },
  "paths": {
...
    "/upload": {
      "post": {
        "summary": "Upload Model",
        "operationId": "upload_model_upload_post",
        "parameters": [
          {
            "name": "syringe_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "left",
              "title": "Syringe Mode"
            }
          },
          {
            "name": "nozzle_diameter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nozzle Diameter"
            }
          },
          {
            "name": "syringe_diameter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Syringe Diameter"
            }
          },
          {
            "name": "layer_height",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer Height"
            }
          },
          {
            "name": "pressurize_mm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pressurize Mm"
            }
          },
          {
            "name": "flow_multiplier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flow Multiplier"
            }
          },
          {
            "name": "travel_retract_multiplier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Travel Retract Multiplier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_model_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload/gcode": {
      "post": {
        "summary": "Upload Gcode",
        "operationId": "upload_gcode_upload_gcode_post",
        "parameters": [
          {
            "name": "syringe_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "left",
              "title": "Syringe Mode"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_gcode_upload_gcode_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/gcode/send": {
      "post": {
        "summary": "Send Gcode",
        "description": "Send a raw G-code line to the printer and return its response.",
        "operationId": "send_gcode_gcode_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GcodeSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GcodeSendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gcode/log": {
      "get": {
        "summary": "Get Serial Log",
        "description": "Return the most recent serial log entries.",
        "operationId": "get_serial_log_gcode_log_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SerialLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_gcode_upload_gcode_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_gcode_upload_gcode_post"
      },
      "Body_upload_model_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_model_upload_post"
      },
...
      "GcodeSendRequest": {
        "properties": {
          "line": {
            "type": "string",
            "title": "Line"
          }
        },
        "type": "object",
        "required": [
          "line"
        ],
        "title": "GcodeSendRequest"
      },
      "GcodeSendResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          },
          "response": {
            "type": "string",
            "title": "Response"
          }
        },
        "type": "object",
        "required": [
          "status",
          "response"
        ],
        "title": "GcodeSendResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SerialLogEntry": {
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "direction": {
            "enum": [
              "sent",
              "received"
            ],
            "title": "Direction",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "line_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Line Number"
          }
        },
        "required": [
          "timestamp",
          "direction",
          "content"
        ],
        "title": "SerialLogEntry",
        "type": "object"
      },
      "SerialLogResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/SerialLogEntry"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "entries"
        ],
        "title": "SerialLogResponse"
      },
...
      "UploadResult": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ready",
            "title": "Status"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "lines_total": {
            "type": "integer",
            "title": "Lines Total"
          },
          "time_estimate_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Estimate S"
          },
          "feed_log_entries": {
            "type": "integer",
            "title": "Feed Log Entries"
          },
          "preview_lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preview Lines"
          }
        },
        "type": "object",
        "required": [
          "status",
          "filename",
          "lines_total",
          "time_estimate_s",
          "feed_log_entries",
          "preview_lines"
        ],
        "title": "UploadResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}